answers/Binary Search/Java/ProgramTest.java
import static org.junit.Assert.assertTrue;
import org.junit.Test;
//public

// This file is initialized with a code version of this
// question's sample test case. Feel free to add, edit,
// or remove test cases in this file as you see fit!



//paste below



// This file is initialized with a code version of this
// question's sample test case. Feel free to add, edit,
// or remove test cases in this file as you see fit!

public class ProgramTest {
  @Test
  public void TestCase1() {
    assertTrue(
        Program.binarySearch(new int[] {0, 1, 21, 33, 45, 45, 61, 71, 72, 73}, 33) == 3);
  }

  @Test
  public void TestCase2() {
    assertTrue(
        Program.binarySearch(new int[] {0, 1, 21, 33, 45, 45, 61, 71, 72, 73}, 333333333) == -1);
  }

  @Test
  public void TestCase333() {
    assertTrue(
        Program.binarySearch(new int[] {0, 1, 21, 33, 45, 45, 61, 71, 72, 73}, 0) == 0);
  }

  

  
}
